+++ CHANGELOG.md
@@ -24,6 +24,9 @@ Changes are grouped as follows
 ### Changed
 - Default value for `limit` in search method is now 100 instead of None to clarify api default behaviour when no limit is passed.
 
+### Added
+- query parameter on asset search.
+
 ## [1.3.4] - 2019-11-07
 
 ### Changed